# go-flow-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Go Flow Gateway API server.",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/line-callback": {
            "get": {
                "description": "Handler the redirect from Line Login and set user session",
                "tags": [
                    "Auth"
                ],
                "summary": "Line Callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code returned from Line Login",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect URL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/line-login": {
            "get": {
                "description": "Redirect to Line Login",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Line Login",
                "responses": {
                    "302": {
                        "description": "Redirect URL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "login information",
                        "name": "LoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "description": "Logout",
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "register information",
                        "name": "RegisterRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesfully registered",
                        "schema": {
                            "$ref": "#/definitions/v1.RegisterResponse"
                        }
                    }
                }
            }
        },
        "/user-profiles": {
            "post": {
                "description": "Create UserProfileRoutes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserProfileRoutes"
                ],
                "summary": "Create UserProfileRoutes",
                "parameters": [
                    {
                        "description": "UserProfileRoutes information",
                        "name": "createUserProfileRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.createUserProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.userProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
        "/user-profiles/{userId}": {
            "get": {
                "description": "Get UserProfileRoutes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserProfileRoutes"
                ],
                "summary": "Get UserProfileRoutes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.userProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
        "/user-uploaded-files": {
            "get": {
                "description": "Get paginated files",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Uploaded File"
                ],
                "summary": "Get paginated files",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "last ID of the last record",
                        "name": "lastID",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.getPaginatedFilesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create user uploaded file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Uploaded File"
                ],
                "summary": "Create user uploaded file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email recipient",
                        "name": "emailRecipient",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.UserUploadedFile": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "emailRecipient": {
                    "description": "The email address of the recipient",
                    "type": "string"
                },
                "emailSent": {
                    "description": "Indicates if the email was sent successfully",
                    "type": "boolean"
                },
                "emailSentAt": {
                    "description": "The timestamp when the email was sent",
                    "type": "string"
                },
                "errorMessage": {
                    "description": "// Error message if the email was not sent successfully",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "v1.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "username": {
                    "type": "string",
                    "example": "username"
                }
            }
        },
        "v1.RegisterRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "displayName": {
                    "type": "string",
                    "example": "billyang"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "username": {
                    "type": "string",
                    "example": "useraname"
                }
            }
        },
        "v1.RegisterResponse": {
            "type": "object",
            "properties": {
                "userID": {
                    "type": "string",
                    "example": "userID"
                }
            }
        },
        "v1.createUserProfileRequest": {
            "type": "object",
            "required": [
                "displayName",
                "pictureUrl"
            ],
            "properties": {
                "displayName": {
                    "type": "string",
                    "example": "John Doe"
                },
                "pictureUrl": {
                    "type": "string",
                    "example": "https://example.com/picture.jpg"
                }
            }
        },
        "v1.errorResponse": {
            "type": "object",
            "properties": {
                "fieldErrs": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "v1.getPaginatedFilesResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.UserUploadedFile"
                    }
                },
                "totalRecords": {
                    "type": "integer"
                }
            }
        },
        "v1.userProfileResponse": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string",
                    "example": "John Doe"
                },
                "pictureUrl": {
                    "type": "string",
                    "example": "https://example.com/picture.jpg"
                }
            }
        }
    }
}